backend/requirements.txt
Flask~=3.0.2
Flask-Cors~=4.0.0
pandas~=2.2.3
gunicorn==23.0
python-dotenv==1.0.1
